[#59] Test coverage for log command
